requirements.txt
flask==3.0.3

langchain==0.2.14
langchain-community==0.2.12
langchain-ollama==0.1.3
langchain-text-splitters==0.2.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==4.3.1
unstructured==0.15.9

python-dotenv==1.0.1

